splitting the unified push reciever logic to a testable delegate
